app/requirements.txt
# FastAPI and Core
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# OAuth
authlib==1.3.0
httpx==0.26.0
itsdangerous==2.1.2

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Security & Encryption
cryptography==42.0.2
pyotp==2.9.0

# OTP & SMS
twilio==9.9.1
redis==7.1.0

# Async HTTP
aiohttp==3.9.3

# WebSockets
websockets==12.0

# Utilities
python-dateutil==2.8.2

# CORS and additional production dependencies
starlette==0.36.3
gunicorn==21.2.0